Removed code that depends on bridged network in mkdiff.sh
Original code attempted to set a static IP for the diff VM, however the
networks are not bridged, so dnsmasq on the host instead supplies IP and
dns via dhcp.  Subsequent lines in mkdiff.sh which attempt to change diff
vm via ssh would fail.  Removed all this, as its intent was to configure
networking and ssh, which is redundant.  Upshot is that this is much faster.